chore: strip out UploadVerifier
